Extract Response subtypes to their own class files
